fix(crash_reporting): enhance crash context logging and memory safety
- Improved crash report filename generation by using localtime_r for thread safety and added checks to prevent buffer overflows and formatting failures.
- Enhanced error handling during file operations to ensure proper resource management and prevent potential crashes.
- Updated timestamp type in CollisionEvent struct from int to uint64_t for better precision and consistency across systems.
- Refactored shader and vertex buffer management to replace NULL with nullptr for improved C++ compatibility.
- Streamlined Vulkan job management by using structured bindings for better readability and maintainability.

+++ src/common/ecs_systems.cpp
@@ -118,7 +118,7 @@ struct CollisionContactResultCallback : public btCollisionWorld::ContactResultCa
 			event.normal[1] = cp.m_normalWorldOnB.y();
 			event.normal[2] = cp.m_normalWorldOnB.z();
 			event.impulse = cp.getAppliedImpulse();
-                        event.timestamp = (int)time(NULL); // Use current time as timestamp
+                        event.timestamp = (uint64_t)time(NULL); // Use current time as timestamp
 
 			events.push_back(event);
 		}
